Move helper class to helper folder
